Generate Factory of the Future colours from a pre-defined palette

Generate Factory of the Future colours from a pre-defined palette

Usage

fof_pal(palette = "main", reverse = FALSE, ...)

Arguments

palette

Name of colour palette. See palette_names() for a list of available palettes.

reverse

Whether the order of the colours should be reversed (default FALSE)

...

Other arguments passed to colorRampPalette()

Value

A colour palette

Examples

# To generate a sequence of n colours from the 'main' palette:
n <- 2
fof_pal("main")(n)
#> [1] "#DC5300" "#232D4B"