AITI Theme

theme_aiti() has been renamed theme_fof()

Usage

theme_aiti(
  base_size = 12,
  colour = "Soft Black",
  legend = "none",
  markdown = FALSE,
  flipped = FALSE,
  legacy = FALSE
)

Arguments

base_size

The base size of text elements of the plot. (default 12)

colour

The background colour of the plot.

legend

The position of the legend (default no legend)

markdown

Whether to use markdown formatting for plot titles (default FALSE)

flipped

TRUE to flip the y-axis guide lines to show on the x-axis instead.

legacy

Whether to use legacy fonts

Value

ggplot2 theme

Details

Theme for plots in AITI publications, reports, Shiny Apps.